+++ tests/testFiles/testInput.txt
@@ -1,0 +1,1 @@
+This is secret. Message about "_" symbol!
